feat: baseInterface.js
- ai21
- deepinfra
- fireworks ai
- friendli ai
- hugging face
- mistral ai
- nvidia
